+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+
+## RELEASE 1.6.14 - 2017-07-05
 ### Added
 - Filters - Add the before x hours operator.